+++ src/main/java/com/meoguri/linkocean/LinkoceanApplication.java
@@ -2,7 +2,6 @@ package com.meoguri.linkocean;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.flyway.FlywayMigrationStrategy;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.session.jdbc.config.annotation.web.http.EnableJdbcHttpSession;
 
@@ -15,13 +14,4 @@ public class LinkoceanApplication {
 		SpringApplication.run(LinkoceanApplication.class, args);
 	}
 
-	// @Bean
-	public FlywayMigrationStrategy flywayMigrationStrategy() {
-		return flyway -> {
-			// flyway.clean();
-			// flyway.repair();
-			// flyway.migrate();
-			flyway.validate();
-		};
-	}
 }
