+++ src/main/java/com/meoguri/linkocean/infrastructure/s3/S3Uploader.java
@@ -5,8 +5,6 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.Optional;
 
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.amazonaws.services.s3.AmazonS3Client;
@@ -18,13 +16,11 @@ import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @RequiredArgsConstructor
-@Component
 public class S3Uploader {
 
 	private final AmazonS3Client amazonS3Client;
 
-	@Value("${cloud.aws.s3.bucket}")
-	private String bucket;
+	private final String bucket;
 
 	public String upload(MultipartFile multipartFile, String dirName) throws IOException {
 		File uploadFile = convert(multipartFile)
